moved kpfcc history code out of history.py
