+++ config_template.ini
@@ -31,7 +31,7 @@ filler_file = fillers.csv
 
 [semester]
 # Size of one slot, in minutes, this must be comparable to the shortest visit + overhead time
-slot_size = 2
+slot_size = 3
 # Run a weather loss model in the autoscheduler
 run_weather_loss = False
 # Specify a random seed. In the paper for weather testing, we use 24 through 33
